backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
httpx
python-dotenv
pydantic
greenlet
